requirements.txt
pymonocypher ~= 4.0.2
